// interfaces/itinerariesCIty.ts
// Generated by https://quicktype.io

export interface ItinerariesCityResponse {
  success: boolean;
  respuesta: Itinerary[];
}

export interface Itinerary {
  likes: number;
  hashtags: string[];
  usersLiked: string[];
  _id: string;
  title: string;
  authorName: string;
  authorPicture: string;
  price: number;
  duration: number;
  cityId: CityID;
  comments: Comment[];
  __v: number;
}

export interface CityID {
  _id: string;
  city: string;
  country: string;
  src: string;
  description: string;
  __v: number;
}

export interface Comment {
  _id: string;
  firstName: string;
  userPicture: string;
  comment: string;
}